major changes to ncep stage IV data grabbing. they switched from grib1 to grib2 files thus making unzipping unncecessary. made a ton of changes to static agwx maps and also the panel applications as well including extra observation plotting on cumulative sum plots

+++ cema/agriculture/deosAG_pointsToGrid.py
@@ -50,7 +50,8 @@ own_cmap1 = mpl.colors.LinearSegmentedColormap.from_list( 'own2', [startcolor, m
 # remove the last 3 days of data - this protects against the scenario where we run before all data is placed in DEOS jsons previously
 rm_dates = pd.date_range(date.today() - timedelta(days=3), date.today() - timedelta(days=1))
 for rd in rm_dates:
-    os.system("/bin/rm " + outPathNC + str(rd.year) + "/" + str("DEOS_agri_" + "{:04d}".format(rd.year) + "{:02d}".format(rd.month) + "{:02d}".format(rd.day) + ".nc"))
+    if os.path.isfile(outPathNC + "/" + str(rd.year) + "/" + str("DEOS_agri_" + "{:04d}".format(rd.year) + "{:02d}".format(rd.month) + "{:02d}".format(rd.day) + ".nc")) == True:
+        os.system("/bin/rm " + outPathNC + str(rd.year) + "/" + str("DEOS_agri_" + "{:04d}".format(rd.year) + "{:02d}".format(rd.month) + "{:02d}".format(rd.day) + ".nc"))
 
 # grab data from json files
 deos_data = pd.read_json("http://128.175.28.202/deos_json/map_data2.json")
@@ -98,7 +99,7 @@ fancyDict = dict(zip(list(nameDict.values()), ['Kelvin', 'Kelvin', 'Kelvin', ' '
 # create a dictionary for months
 monthDict = dict(zip([1,2,3,4,5,6,7,8,9,10,11,12], ['Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun', 'Jul', 'Aug', 'Sep', 'Oct', 'Nov', 'Dec']))
 
-daysback = range(1,5)
+daysback = range(8,0,-1)
 for dy in daysback:
     nowtime = datetime.utcnow() - timedelta(days=dy)
     daytime = str("{:04d}".format(nowtime.year) + "-" + "{:02d}".format(nowtime.month) + "-" + "{:02d}".format(nowtime.day))
@@ -125,7 +126,7 @@ for dy in daysback:
                         try:
                             agJson = pd.read_json(stat_path)
                             # use this for when we are real-time et.append(int(float(et_data[rev_station_dict[key]][str(str(nowtime.year) + "-" + str("{0:0=2d}".format(nowtime.month)) + "-" + str("{0:0=2d}".format(nowtime.day)))]['Reference Evapotrans.']['Value'])))
-                            varData.append(round(float(agJson[rev_station_dict[key]][str(daytime + ' 00:00:00')][var]['Value']),4))
+                            varData.append(round(float(agJson[rev_station_dict[key]][str(daytime + ' 00:00')][var]['Value']),4))
                             lats.append(loc_deos[rev_station_dict[key]]['latitude'])
                             lons.append(loc_deos[rev_station_dict[key]]['longitude'])
                             workKey.append(str(key))
